scraper/requirements.txt
langchain>=0.0.300
scrapegraph-sdk>=0.0.1
requests>=2.28
beautifulsoup4>=4.12
python-dotenv>=1.0
pytest>=7.0
psycopg2-binary>=2.9.0
